refact: abstract suggestSlug()
