+++ lib/collective.lib.js
@@ -1,4 +1,5 @@
-import { get } from 'lodash';
+import { get, trim } from 'lodash';
+import slugify from 'slugify';
 
 import {
   CollectiveCategory,
@@ -63,4 +64,14 @@ export const getCollectiveTypeForUrl = collective => {
 
 export const hostIsTaxDeductibeInTheUs = host => {
   return get(host, 'settings.taxDeductibleDonations');
+};
+
+export const suggestSlug = value => {
+  const slugOptions = {
+    replacement: '-',
+    lower: true,
+    strict: true,
+  };
+
+  return trim(slugify(value, slugOptions), '-');
 };
